Capitalize only makes the first letter capital
Use our own word capitalization function that only make the first letter of
each word capital without changing the other letters in the word.

For example, using string.capwords, TransJakarta and KWK will be converted to
Transjakarta and Kwk respectively. This is not desirable since they are already
correct.

Issue #88